X-Code Mac 449183d45b9f23170001bc8c58ff18ea
a4cadedf158183f31eafe0b13af67cb1062dca6acfb72ae4f180d9bc73db7d7b

+++ Code Template/Code_Template/Headers/functions.h
@@ -181,7 +181,7 @@ uint64_t unsigned_ripple_carry_adder(uint64_t a, uint64_t b)
     return sum | carryin;
 }
 
-long double ripple_multiply(uint64_t x, uint64_t y)
+uint64_t ripple_multiply(uint64_t x, uint64_t y)
 {
     uint64_t sum = 0;
     while(x)
